6/26/2022_1:08PM
Currently working on the Drawing module. There are some bugs in tasks 59 and 70, they don't happen all the time. Additionally there is a null reference exception every time 'draw bond' is called. With the technology I have at hand, I cannot identify the sources of this problems, so I will move forward to the next part.
Next task is to ensure the clear system works and that users can restart the drawing module.